+++ packages/amplify-codegen/src/commands/models.js
@@ -20,7 +20,7 @@ async function generateModels(context) {
 
   const allApiResources = await context.amplify.getResourceStatus('api');
   const apiResource = allApiResources.allResources.find(
-    resource => resource.service === 'AppSync' && resource.providerPlugin === 'awscloudformation'
+    resource => resource.service === 'AppSync' && resource.providerPlugin === 'awscloudformation',
   );
 
   if (!apiResource) {
@@ -105,7 +105,9 @@ function getModelOutputPath(context) {
     case 'javascript':
       return 'src';
     case 'android':
-      return 'app/src/main/java/';
+      return projectConfig.android && projectConfig.android.config && projectConfig.android.config.ResDir
+        ? path.normalize(path.join(projectConfig.android.config.ResDir, '..', 'java'))
+        : path.join('app', 'src', 'main', 'java');
     case 'ios':
       return 'amplify/generated/models';
     default: