+++ packages/amplify-dotnet-function-template-provider/src/utils/constants.ts
@@ -1,1 +1,4 @@
-export const templateRoot = `${__dirname}/../../resources`;
+const templateRoot = `${__dirname}/../../resources`;
+const commonFiles = ['global.json', '.gitignore.ejs'];
+
+export { templateRoot, commonFiles };