+++ packages/amplify-e2e-tests/src/__tests__/predictions.test.ts
@@ -1,7 +1,8 @@
-import { initJSProjectWithProfile, deleteProject, amplifyPushAuth } from '../init';
-import { createNewProjectDir, deleteProjectDir, getAWSExports, getCollection } from '../utils';
-import { addConvert, addInterpret, addIdentifyCollection } from '../categories/predictions';
-import { addAuthWithDefault } from '../categories/auth';
+import { initJSProjectWithProfile, deleteProject, amplifyPushAuth } from 'amplify-e2e-core';
+import { createNewProjectDir, deleteProjectDir, getCollection } from 'amplify-e2e-core';
+import { addConvert, addInterpret, addIdentifyCollection } from 'amplify-e2e-core';
+import { addAuthWithDefault } from 'amplify-e2e-core';
+import { getAWSExports } from '../aws-exports/awsExports';
 
 describe('amplify add predictions', () => {
   let projRoot: string;