+++ packages/amplify-e2e-tests/src/amplify-app-helpers/amplify-app-setup.ts
@@ -1,7 +1,6 @@
-import { nspawn as spawn, KEY_DOWN_ARROW } from 'amplify-e2e-core';
+import { nspawn as spawn, KEY_DOWN_ARROW, isCI } from 'amplify-e2e-core';
 import * as fs from 'fs-extra';
 import * as path from 'path';
-import { isCI } from '../utils';
 
 const npm = /^win/.test(process.platform) ? 'npm.cmd' : 'npm';
 const amplifyAppBinPath = path.join(__dirname, '..', '..', '..', 'amplify-app', 'bin', 'amplify-app');