+++ packages/amplify-e2e-tests/src/codegen/add.ts
@@ -1,5 +1,4 @@
-import { nspawn as spawn } from 'amplify-e2e-core';
-import { getCLIPath } from '../utils';
+import { nspawn as spawn, getCLIPath } from 'amplify-e2e-core';
 
 export function addCodegen(cwd: string, settings: any) {
   return new Promise((resolve, reject) => {