+++ packages/graphql-connection-transformer/src/resources.ts
@@ -185,8 +185,7 @@ export class ResourceFactory {
     sortKeyInfo?: { fieldName: string; attributeType: 'S' | 'B' | 'N' },
     limit?: number,
   ) {
-    const defaultPageLimit = 10;
-    const pageLimit = limit || defaultPageLimit;
+    const pageLimit = limit || ResourceConstants.DEFAULT_PAGE_LIMIT;
     const setup: Expression[] = [
       set(ref('limit'), ref(`util.defaultIfNull($context.args.limit, ${pageLimit})`)),
       set(
@@ -312,8 +311,7 @@ export class ResourceFactory {
     indexName: string,
     limit?: number,
   ) {
-    const defaultPageLimit = 10;
-    const pageLimit = limit || defaultPageLimit;
+    const pageLimit = limit || ResourceConstants.DEFAULT_PAGE_LIMIT;
     const setup: Expression[] = [
       set(ref('limit'), ref(`util.defaultIfNull($context.args.limit, ${pageLimit})`)),
       set(ref('query'), this.makeExpression(keySchema, connectionAttributes)),