+++ packages/amplify-console-integration-tests/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.3](https://github.com/aws-amplify/amplify-console-integration-tests/compare/amplify-console-integration-tests@1.2.2...amplify-console-integration-tests@1.2.3) (2020-04-23)
+
+**Note:** Version bump only for package amplify-console-integration-tests
+
+
+
+
+
 ## 1.2.2 (2020-04-06)
 
 **Note:** Version bump only for package amplify-console-integration-tests